app/requirements.txt
Django==3.1.1
djangorestframework==3.12.1
psycopg2==2.8.6
gunicorn==20.0.4
